Study_Flask/standard_flask_survey/pyproject.toml
[tool.poetry]
name = "flask-quiz"
version = "0.1.0"
description = ""
authors = ["kwanyeol <djm030@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.2"
flask-sqlalchemy = "^3.1.1"
flask-mysqldb = "^2.0.0"
flask-migrate = "^4.0.5"
pandas = "^2.2.1"
plotly = "^5.19.0"
gunicorn = "^21.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
